[TIDY] Move RPC methods to EthereumRPCProtocol, so those can be shared with ZKSync
