+++ web3swift/src/Client/BaseEthereumClient.swift
@@ -41,259 +41,6 @@ open class BaseEthereumClient: EthereumClientProtocol {
         }
     }
 
-    public func net_version() async throws -> EthereumNetwork {
-        let emptyParams: [Bool] = []
-        do {
-            let data = try await networkProvider.send(method: "net_version", params: emptyParams, receive: String.self)
-
-            if let resString = data as? String {
-                let network = EthereumNetwork.fromString(resString)
-                return network
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_gasPrice() async throws -> BigUInt {
-        let emptyParams: [Bool] = []
-
-        do {
-            let data = try await networkProvider.send(method: "eth_gasPrice", params: emptyParams, receive: String.self)
-            if let hexString = data as? String, let bigUInt = BigUInt(hex: hexString) {
-                return bigUInt
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_blockNumber() async throws -> Int {
-        let emptyParams: [Bool] = []
-
-        do {
-            let data = try await networkProvider.send(method: "eth_blockNumber", params: emptyParams, receive: String.self)
-            if let hexString = data as? String {
-                if let integerValue = Int(hex: hexString) {
-                    return integerValue
-                } else {
-                    throw EthereumClientError.decodeIssue
-                }
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_getBalance(address: EthereumAddress, block: EthereumBlock) async throws -> BigUInt {
-        do {
-            let data = try await networkProvider.send(method: "eth_getBalance", params: [address.asString(), block.stringValue], receive: String.self)
-            if let resString = data as? String, let balanceInt = BigUInt(hex: resString.web3.noHexPrefix) {
-                return balanceInt
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_getCode(address: EthereumAddress, block: EthereumBlock = .Latest) async throws -> String {
-        do {
-            let data = try await networkProvider.send(method: "eth_getCode", params: [address.asString(), block.stringValue], receive: String.self)
-            if let resDataString = data as? String {
-                return resDataString
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_estimateGas(_ transaction: EthereumTransaction) async throws -> BigUInt {
-        struct CallParams: Encodable {
-            let from: String?
-            let to: String
-            let value: String?
-            let data: String?
-
-            enum TransactionCodingKeys: String, CodingKey {
-                case from
-                case to
-                case value
-                case data
-            }
-
-            func encode(to encoder: Encoder) throws {
-                var container = encoder.unkeyedContainer()
-                var nested = container.nestedContainer(keyedBy: TransactionCodingKeys.self)
-                if let from = from {
-                    try nested.encode(from, forKey: .from)
-                }
-                try nested.encode(to, forKey: .to)
-
-                let jsonRPCAmount: (String) -> String = { amount in
-                    amount == "0x00" ? "0x0" : amount
-                }
-
-                if let value = value.map(jsonRPCAmount) {
-                    try nested.encode(value, forKey: .value)
-                }
-                if let data = data {
-                    try nested.encode(data, forKey: .data)
-                }
-            }
-        }
-
-        let value: BigUInt?
-        if let txValue = transaction.value, txValue > .zero {
-            value = txValue
-        } else {
-            value = nil
-        }
-
-        let params = CallParams(
-            from: transaction.from?.asString(),
-            to: transaction.to.asString(),
-            value: value?.web3.hexStringNoLeadingZeroes,
-            data: transaction.data?.web3.hexString
-        )
-
-        do {
-            let data = try await networkProvider.send(method: "eth_estimateGas", params: params, receive: String.self)
-            if let gasHex = data as? String, let gas = BigUInt(hex: gasHex) {
-                return gas
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_sendRawTransaction(_ transaction: EthereumTransaction, withAccount account: EthereumAccountProtocol) async throws -> String {
-        do {
-            // Inject pending nonce
-            let nonce = try await eth_getTransactionCount(address: account.address, block: .Pending)
-
-            var transaction = transaction
-            transaction.nonce = nonce
-
-            if transaction.chainId == nil, let network = network {
-                transaction.chainId = network.intValue
-            }
-
-            guard let _ = transaction.chainId, let signedTx = (try? account.sign(transaction: transaction)), let transactionHex = signedTx.raw?.web3.hexString else {
-                throw EthereumClientError.encodeIssue
-            }
-
-            let data = try await networkProvider.send(method: "eth_sendRawTransaction", params: [transactionHex], receive: String.self)
-            if let resDataString = data as? String {
-                return resDataString
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_getTransaction(byHash txHash: String) async throws -> EthereumTransaction {
-        do {
-            let data = try await networkProvider.send(method: "eth_getTransactionByHash", params: [txHash], receive: EthereumTransaction.self)
-            if let transaction = data as? EthereumTransaction {
-                return transaction
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_getTransactionReceipt(txHash: String) async throws -> EthereumTransactionReceipt {
-        do {
-            let data = try await networkProvider.send(method: "eth_getTransactionReceipt", params: [txHash], receive: EthereumTransactionReceipt.self)
-            if let receipt = data as? EthereumTransactionReceipt {
-                return receipt
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
-    public func eth_getLogs(addresses: [EthereumAddress]?, topics: [String?]?, fromBlock from: EthereumBlock = .Earliest, toBlock to: EthereumBlock = .Latest) async throws -> [EthereumLog] {
-        try await RecursiveLogCollector(ethClient: self).getAllLogs(addresses: addresses, topics: topics.map(Topics.plain), from: from, to: to)
-    }
-
-    public func eth_getLogs(addresses: [EthereumAddress]?, orTopics topics: [[String]?]?, fromBlock from: EthereumBlock = .Earliest, toBlock to: EthereumBlock = .Latest) async throws -> [EthereumLog] {
-        try await RecursiveLogCollector(ethClient: self).getAllLogs(addresses: addresses, topics: topics.map(Topics.composed), from: from, to: to)
-    }
-
-    public func getLogs(addresses: [EthereumAddress]?, topics: Topics?, fromBlock: EthereumBlock, toBlock: EthereumBlock) async throws -> [EthereumLog] {
-        struct CallParams: Encodable {
-            var fromBlock: String
-            var toBlock: String
-            let address: [EthereumAddress]?
-            let topics: Topics?
-        }
-
-        let params = CallParams(fromBlock: fromBlock.stringValue, toBlock: toBlock.stringValue, address: addresses, topics: topics)
-
-        do {
-            let data = try await networkProvider.send(method: "eth_getLogs", params: [params], receive: [EthereumLog].self)
-
-            if let logs = data as? [EthereumLog] {
-                return logs
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            if let error = error as? JSONRPCError,
-               case let .executionError(innerError) = error,
-               innerError.error.code == JSONRPCErrorCode.tooManyResults {
-                throw EthereumClientError.tooManyResults
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        }
-    }
-
-    public func eth_getBlockByNumber(_ block: EthereumBlock) async throws -> EthereumBlockInfo {
-        struct CallParams: Encodable {
-            let block: EthereumBlock
-            let fullTransactions: Bool
-
-            func encode(to encoder: Encoder) throws {
-                var container = encoder.unkeyedContainer()
-                try container.encode(block.stringValue)
-                try container.encode(fullTransactions)
-            }
-        }
-
-        let params = CallParams(block: block, fullTransactions: false)
-
-        do {
-            let data = try await networkProvider.send(method: "eth_getBlockByNumber", params: params, receive: EthereumBlockInfo.self)
-            if let blockData = data as? EthereumBlockInfo {
-                return blockData
-            } else {
-                throw EthereumClientError.unexpectedReturnValue
-            }
-        } catch {
-            throw failureHandler(error)
-        }
-    }
-
     private func fetchNetwork() async -> EthereumNetwork? {
         do {
             return try await net_version()
